Fixed async tasks and added UI test project to automate the smoke testing of all controllers
