+++ solution/WebApplication/WebApplication/Controllers/ScheduleInstanceController.cs
@@ -230,7 +230,7 @@ namespace WebApplication.Controllers
         }
 
         [ChecksUserAccess]
-        public ActionResult GetGridData()
+        public async Task<ActionResult> GetGridData()
         {
             try
             {
@@ -296,9 +296,9 @@ namespace WebApplication.Controllers
                     .Include(t => t.ScheduleMaster).AsNoTracking();
 
                 //total number of rows count     
-                recordsTotal = modelDataAll.Count();
+                recordsTotal = await modelDataAll.CountAsync();
                 //Paging     
-                var data = modelDataAll.Skip(skip).Take(pageSize).ToList();
+                var data = await modelDataAll.Skip(skip).Take(pageSize).ToListAsync();
                 //Returning Json Data    
                 var jserl = new JsonSerializerSettings
                 {
