+++ solution/WebApplication/WebApplication/Controllers/TaskMasterController.cs
@@ -258,7 +258,7 @@ namespace WebApplication.Controllers
         }
 
         [ChecksUserAccess]
-        public ActionResult GetGridData()
+        public async Task<ActionResult> GetGridData()
         {
             try
             {
@@ -342,11 +342,11 @@ namespace WebApplication.Controllers
 
 
                 //total number of rows count     
-                recordsTotal = modelDataAll.Count();
+                recordsTotal = await modelDataAll.CountAsync();
 
 
                 //Paging               
-                var data = modelDataAll.Skip(skip).Take(pageSize).ToList();
+                var data = await modelDataAll.Skip(skip).Take(pageSize).ToListAsync();
 
 
 
@@ -383,7 +383,7 @@ namespace WebApplication.Controllers
                 tm.TaskMasterName = tm.TaskMasterName + " Copy";
                 _context.Add(tm);
             }
-            _context.SaveChanges();
+            await _context.SaveChangesAsync();
 
             //TODO: Add Error Handling
             return new OkObjectResult(new { });
@@ -403,7 +403,7 @@ namespace WebApplication.Controllers
 
                 ti.ActiveYn = Status;
             }
-            _context.SaveChanges();
+            await _context.SaveChangesAsync();
 
             //TODO: Add Error Handling
             return new OkObjectResult(new { });
