+++ solution/WebApplication/WebApplication/Controllers/TaskMasterWaterMarkController.cs
@@ -229,7 +229,7 @@ namespace WebApplication.Controllers
         }
 
         [ChecksUserAccess]
-        public ActionResult GetGridData()
+        public async Task<ActionResult> GetGridData()
         {
             try
             {
@@ -300,11 +300,11 @@ namespace WebApplication.Controllers
 
 
                 //total number of rows count     
-                recordsTotal = modelDataAll.Count();
+                recordsTotal = await modelDataAll.CountAsync();
 
 
                 //Paging               
-                var data = modelDataAll.Skip(skip).Take(pageSize).ToList();
+                var data = await modelDataAll.Skip(skip).Take(pageSize).ToListAsync();
 
 
 
@@ -340,7 +340,7 @@ namespace WebApplication.Controllers
 
                 ti.ActiveYn = Status;
             }
-            _context.SaveChanges();
+            await _context.SaveChangesAsync();
 
             //TODO: Add Error Handling
             return new OkObjectResult(new { });
