+++ solution/WebApplication/WebApplication/Controllers/TaskTypeController.cs
@@ -219,7 +219,7 @@ namespace WebApplication.Controllers
             return new OkObjectResult(JsonConvert.SerializeObject(GridCols()));
         }
 
-        public ActionResult GetGridData()
+        public async Task<ActionResult> GetGridData()
         {
             try
             {
@@ -252,9 +252,9 @@ namespace WebApplication.Controllers
                 }
 
                 //total number of rows count     
-                recordsTotal = modelDataAll.Count();
+                recordsTotal = await modelDataAll.CountAsync();
                 //Paging     
-                var data = modelDataAll.Skip(skip).Take(pageSize).ToList();
+                var data = await modelDataAll.Skip(skip).Take(pageSize).ToListAsync();
                 //Returning Json Data    
                 return new OkObjectResult(JsonConvert.SerializeObject(new { draw = draw, recordsFiltered = recordsTotal, recordsTotal = recordsTotal, data = data }, new Newtonsoft.Json.Converters.StringEnumConverter()));
